Refactor exceptions
Many of the exception classes were refurbished, and the unit tests for the parametric distributions had to be accordingly updated. Inference unit tests are running, but a couple of them that require exception handling are failing -- will need to fix those next. Almost all in interface/ is PEP8 compliant, one more file to go before moving on to the next module.