Bug fix in event series parsing
Fixed nasty bug related to providing an empty mutable as the default argument to an __init__ method. Also fixed stochastic map example file for two regions, as well as the paleogeographic data for that same example.

Added a new barrier appearance and disappearance classes that are now included in the event series. The barrier event is probably not necessary, but the barrier disappearance one is, because it tells us if the range became stable again.

Need to next write unit tests for the simple 2 region stochastic map / event series parsing. Then write a new method for truncation of that.